Sixtyfps (#3)
Added alternative UI framework implementation: sixtyfps. Now, to build one of the features much be enabled: qml_ui or sixtyfps_ui.

+++ src/game/domino.rs
@@ -1,7 +1,7 @@
 use crate::game::board::Position;
 use std::fmt;
 
-pub type Id = u8;
+pub type Id = u32;
 pub type Value = u8;
 
 #[derive(Copy, Clone, Debug, Eq, Hash, PartialEq)]
